+++ lib/src/brands/reddit.dart
@@ -1,20 +1,20 @@
 import 'package:eo_color/eo_color.dart';
 
-/// Reddint's official brand colors
+/// Reddit's official brand colors
 ///
 /// - Black;
 /// - Orange;
 /// - White.
 ///
 /// See also:
-/// - [reddint colors](https://brandpalettes.com/reddit-color-codes/)
-class Reddint extends PaletteRGB {
+/// - [reddit colors](https://brandpalettes.com/reddit-color-codes/)
+class Reddit extends PaletteRGB {
   /// Black #000000.
-  const Reddint.black() : super.black();
+  const Reddit.black() : super.black();
 
   /// Orange #FF4500.
-  const Reddint.orange() : super.opaque(0xFF4500);
+  const Reddit.orange() : super.opaque(0xFF4500);
 
   /// White #FFFFFF.
-  const Reddint.white() : super.white();
+  const Reddit.white() : super.white();
 }